feat: split the insight tool
