+++ ee/hogai/graph/root/tools/__init__.py
@@ -1,4 +1,5 @@
-from .legacy import create_and_query_insight, create_dashboard, session_summarization
+from .create_and_query_insight import CreateAndQueryInsightTool, CreateAndQueryInsightToolArgs
+from .legacy import create_dashboard, session_summarization
 from .navigate import NavigateTool, NavigateToolArgs
 from .read_data import ReadDataTool, ReadDataToolArgs
 from .read_taxonomy import ReadTaxonomyTool
@@ -15,7 +16,8 @@ __all__ = [
     "TodoWriteToolArgs",
     "NavigateTool",
     "NavigateToolArgs",
-    "create_and_query_insight",
     "session_summarization",
     "create_dashboard",
+    "CreateAndQueryInsightTool",
+    "CreateAndQueryInsightToolArgs",
 ]
